修改融资  新增 修改 删除 搞定

+++ loan-service/src/main/java/com/zsy/loan/service/system/IConstantFactory.java
@@ -269,7 +269,7 @@ public interface IConstantFactory {
    * @param type
    * @return
    */
-  String getInvestTypeName(Long type);
+  String getInvestTypeName(Integer type);
 
   /**
    * 获取回款状态名称
